chore: Move test helpers to directory
