+++ tests/setup.ts
@@ -1,5 +1,5 @@
 import { afterAll } from "vitest";
-import { generateAllTests } from "./_type_testing";
+import { generateAllTests } from "./_helpers/typechecks";
 
 afterAll(() => {
 	generateAllTests();
