+++ tests/utils/tree.test.ts
@@ -1,4 +1,4 @@
-import { it } from "^/tests/_helpers";
+import { it } from "^/tests/_helpers/vitest";
 import { expect } from "vitest";
 import { TreeNode, type TreeSource, makeTree } from "~/utils/tree";
